core/services/network/machine_discovery/requirements.txt
redis == 3.3.11
python-nmap == 0.6.1
